renamed simpletenantusage to usage
