+++ openstack/compute/v2/extensions/usage/results.go
@@ -1,4 +1,4 @@
-package simpletenantusage
+package usage
 
 import (
 	"encoding/json"
