+++ openstack/compute/v2/extensions/usage/testing/requests_test.go
@@ -3,19 +3,19 @@ package testing
 import (
 	"testing"
 
-	"github.com/gophercloud/gophercloud/openstack/compute/v2/extensions/simpletenantusage"
+	"github.com/gophercloud/gophercloud/openstack/compute/v2/extensions/usage"
 	th "github.com/gophercloud/gophercloud/testhelper"
 	"github.com/gophercloud/gophercloud/testhelper/client"
 )
 
 func TestGetTenant(t *testing.T) {
-	var getOpts simpletenantusage.SingleTenantOpts
+	var getOpts usage.SingleTenantOpts
 	th.SetupHTTP()
 	defer th.TeardownHTTP()
 	HandleGetSingleTenantSuccessfully(t)
-	page, err := simpletenantusage.SingleTenant(client.ServiceClient(), FirstTenantID, getOpts).AllPages()
+	page, err := usage.SingleTenant(client.ServiceClient(), FirstTenantID, getOpts).AllPages()
 	th.AssertNoErr(t, err)
-	actual, err := simpletenantusage.ExtractSingleTenant(page)
+	actual, err := usage.ExtractSingleTenant(page)
 	th.AssertNoErr(t, err)
 	th.CheckDeepEquals(t, &SingleTenantUsageResults, actual)
 }
