+++ openstack/compute/v2/extensions/usage/urls.go
@@ -1,4 +1,4 @@
-package simpletenantusage
+package usage
 
 import "github.com/gophercloud/gophercloud"
